pyproject.toml
[project]
name = "agent-serving-lab"
version = "0.1.0"
description = "Serve deep agents on vLLM for efficient performance and accuracy over long running tasks"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain",
    "langgraph",
    "langchain-aws",
    "zmq>=0.0.0",
    "tavily-python>=0.7.12",
    "deepagents>=0.0.5",
    "langchain-openai>=0.3.35",
    "python-dotenv>=1.1.1",
]
